requirements.txt
numpy==1.19.5
opencv-python-headless==4.5.4.58
scipy==1.7.3
tensorflow==2.7.0
networkx==2.6.3
pygame==2.1.1
imutils==0.5.4
scikit-image==0.19.5
matplotlib==3.5.1
pathlib==1.0.1
Pillow==8.4.0
vidgear==0.3.3
tensorflow==2.7.0
tensorflow-hub==0.12.0
pandas==1.3.5
tflite-model-maker==0.3.5

